chore: separate code and doc generator
