+++ projects/angular-codegen/src/lib/angular-codegen.service.ts
@@ -44,12 +44,6 @@ export class AngularCodeGenService {
 
     return [
       {
-        uri: 'README.md',
-        value: this.renderReadme(data.meta.app),
-        language: 'markdown',
-        kind: 'text'
-      },
-      {
         uri: 'xlayers-routing.module.ts',
         value: this.renderRoutingModule(),
         language: 'typescript',
@@ -130,42 +124,6 @@ export class AngularCodeGenService {
       assetDir: 'assets',
       ...options
     };
-  }
-
-  private renderReadme(name: string) {
-    return `\
-## How to use the ${name} Vue module
-
-1. Download and extract the exported module into your workspace,
-
-2. Option #1: Import eagerly the XlayersModule into your AppModule or other module.
-\`\`\`
-import { XlayersModule } from './xlayers/xlayers.module';
-
-@NgModule({
-  imports: [
-    XlayersModule,
-    ...
-  ],
-})
-export class AppModule {}
-\`\`\`
-
-2. Option #2: Import lazily the XlayersModule routing configuration into your AppModule or other module.
-Make sure your router is setup properly in order to use this option (see: https://angular.io/guide/lazy-loading-ngmodules).
-
-\`\`\`
-import { XlayersRoutingModule } from './xlayers/xlayers-routing.module';
-@NgModule({
-  imports: [
-    XlayersRoutingModule,
-    ...
-  ],
-})
-export class AppModule {}
-\`\`\`
-
-3. Enjoy.`;
   }
 
   private renderRoutingModule() {
