+++ projects/angular-codegen/src/lib/angular-docgen.service.ts
@@ -1,0 +1,53 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class AngularDocGenService {
+  aggreate(data: SketchMSData) {
+    return [
+      {
+        uri: 'README.md',
+        value: this.renderReadme(data.meta.app),
+        language: 'markdown',
+        kind: 'text'
+      }
+    ];
+  }
+
+  private renderReadme(name: string) {
+    return `\
+## How to use the ${name} Vue module
+
+1. Download and extract the exported module into your workspace,
+
+2. Option #1: Import eagerly the XlayersModule into your AppModule or other module.
+\`\`\`
+import { XlayersModule } from './xlayers/xlayers.module';
+
+@NgModule({
+  imports: [
+    XlayersModule,
+    ...
+  ],
+})
+export class AppModule {}
+\`\`\`
+
+2. Option #2: Import lazily the XlayersModule routing configuration into your AppModule or other module.
+Make sure your router is setup properly in order to use this option (see: https://angular.io/guide/lazy-loading-ngmodules).
+
+\`\`\`
+import { XlayersRoutingModule } from './xlayers/xlayers-routing.module';
+@NgModule({
+  imports: [
+    XlayersRoutingModule,
+    ...
+  ],
+})
+export class AppModule {}
+\`\`\`
+
+3. Enjoy.`;
+  }
+}
