+++ projects/angular-codegen/src/public_api.ts
@@ -2,5 +2,6 @@
  * Public API Surface of angular-codegen
  */
 
+export * from './lib/angular-docgen.service';
 export * from './lib/angular-codegen.service';
 export * from './lib/angular-codegen.module';
