+++ projects/lit-element-codegen/src/lib/lit-element-codegen.service.ts
@@ -30,14 +30,7 @@ export class LitElementCodeGenService {
     data: SketchMSData,
     options?: WebCodeGenOptions
   ) {
-    return [
-      {
-        uri: 'README.md',
-        value: this.renderReadme(data.meta.app),
-        language: 'markdown',
-        kind: 'text'
-      }
-    ].concat(this.visit(current, data, this.compileOptions(options)));
+    return this.visit(current, data, this.compileOptions(options));
   }
 
   identify(current: SketchMSLayer) {
@@ -106,39 +99,5 @@ export class LitElementCodeGenService {
       assetDir: 'assets',
       ...options
     };
-  }
-
-  private renderReadme(name: string) {
-    return `\
-## How to use the ${name} Web Components built with LitElement
-
-This implementation export the assets as single file web component that can be consumed in the following ways:
-
-\`\`\`html
-<!–– index.html ––>
-<script src="./x-layers-element.js"></script>
-<x-layers-element></x-layers-element>
-\`\`\`
-
-> Needed polyfills must be imported, in most cases you can import it globally or use different strategy. For example:
-
-\`\`\`html
-<!–– index.html ––>
-<!-- Load polyfills; note that "loader" will load these async -->
-<script src="node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js" defer></script>
-
-<!-- Load a custom element definitions in \'waitFor\' and return a promise -->
-<script type="module">
-  WebComponents.waitFor(() => {
-    return import(\'./x-layers-element.js\');
-  });
-</script>
-
-<!-- Use the custom element -->
-<x-layers-element></x-layers-element>
-\`\`\`
-
->  [LitElement website](https://lit-element.polymer-project.org/)
->  For more information about [web components and browser support](https://github.com/WebComponents/webcomponentsjs#browser-support)`;
   }
 }
