+++ projects/lit-element-codegen/src/public_api.ts
@@ -2,5 +2,6 @@
  * Public API Surface of lit-element-codegen
  */
 
+export * from './lib/lit-element-docgen.service';
 export * from './lib/lit-element-codegen.service';
 export * from './lib/lit-element-codegen.module';
