+++ projects/react-codegen/src/lib/react-codegen.service.ts
@@ -30,14 +30,7 @@ export class ReactCodeGenService {
     data: SketchMSData,
     options?: WebCodeGenOptions
   ) {
-    return [
-      {
-        uri: 'README.md',
-        value: this.renderReadme(data.meta.app),
-        language: 'markdown',
-        kind: 'text'
-      }
-    ].concat(this.visit(current, data, this.compileOptions(options)));
+    return this.visit(current, data, this.compileOptions(options));
   }
 
   identify(current: SketchMSLayer) {
@@ -106,24 +99,5 @@ export class ReactCodeGenService {
       assetDir: 'assets',
       ...options
     };
-  }
-
-  private renderReadme(name: string) {
-    return `\
-## How to use the ${name} Vue module
-
-Import and use it with ReactDOM :
-
-\`\`\`javascript
-import ReactDOM from "react-dom";
-import { MyComponent } from "./my-component";
-
-ReactDOM.aggreate(
-  MyComponent,
-  document.getElementById(\'root\')
-);
-\`\`\`
-
->  For more information about [Reactjs](https://reactjs.org/)`;
   }
 }
