+++ projects/react-codegen/src/public_api.ts
@@ -2,5 +2,6 @@
  * Public API Surface of react-codegen
  */
 
+export * from './lib/react-docgen.service';
 export * from './lib/react-codegen.service';
 export * from './lib/react-codegen.module';
