+++ projects/stencil-codegen/src/lib/stencil-docgen.service.ts
@@ -1,0 +1,35 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class StencilDocGenService {
+  aggreate(data: SketchMSData) {
+    return [
+      {
+        uri: 'README.md',
+        value: this.renderReadme(data.meta.app),
+        language: 'markdown',
+        kind: 'text'
+      }
+    ];
+  }
+
+  private renderReadme(name: string) {
+    return `\
+## How to use the ${name} StencilJS Web Components
+
+This implementation export all assets needed to build stenciljs component
+
+Simple use :
+\`\`\`html
+  // index.html
+  <script src="./my-component.js"></script>
+  <my-component></my-component>
+\`\`\`
+
+For more examples how to integrate into your application, view [Framework Integrations](https://stenciljs.com/docs/overview)
+
+>  For more information about [Stenciljs](https://stenciljs.com/)`;
+  }
+}
