+++ projects/stencil-codegen/src/public_api.ts
@@ -2,5 +2,6 @@
  * Public API Surface of stencil-codegen
  */
 
+export * from './lib/stencil-docgen.service';
 export * from './lib/stencil-codegen.service';
 export * from './lib/stencil-codegen.module';
