+++ projects/vue-codegen/src/lib/vue-codegen.service.ts
@@ -30,14 +30,7 @@ export class VueCodeGenService {
     data: SketchMSData,
     options?: WebCodeGenOptions
   ) {
-    return [
-      {
-        uri: 'README.md',
-        value: this.renderReadme(data.meta.app),
-        language: 'markdown',
-        kind: 'text'
-      }
-    ].concat(this.visit(current, data, this.compileOptions(options)));
+    return this.visit(current, data, this.compileOptions(options));
   }
 
   identify(current: SketchMSLayer) {
@@ -106,34 +99,5 @@ export class VueCodeGenService {
       assetDir: 'assets',
       ...options
     };
-  }
-
-  private renderReadme(name: string) {
-    return `\
-## How to use the ${name} Vue module
-
-1. Download and extract the exported module into your workspace,
-
-2. Import the component into your App component or other container.
-\`\`\`
-<template>
-  <div id="app">
-    <MyComponent />
-  </div>
-</template>
-
-<script>
-import MyComponent from \'./components/MyComponent.vue\'
-
-export default {
-  name: \'app\
-  components: {
-    MyComponent
-  }
-}
-</script>
-\`\`\`
-
-3. Enjoy.`;
   }
 }
