+++ projects/vue-codegen/src/lib/vue-docgen.service.ts
@@ -1,0 +1,46 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class VueDocGenService {
+  aggreate(data: SketchMSData) {
+    return [
+      {
+        uri: 'README.md',
+        value: this.renderReadme(data.meta.app),
+        language: 'markdown',
+        kind: 'text'
+      }
+    ];
+  }
+
+  private renderReadme(name: string) {
+    return `\
+## How to use the ${name} Vue module
+
+1. Download and extract the exported module into your workspace,
+
+2. Import the component into your App component or other container.
+\`\`\`
+<template>
+  <div id="app">
+    <MyComponent />
+  </div>
+</template>
+
+<script>
+import MyComponent from \'./components/MyComponent.vue\'
+
+export default {
+  name: \'app\
+  components: {
+    MyComponent
+  }
+}
+</script>
+\`\`\`
+
+3. Enjoy.`;
+  }
+}
