+++ projects/vue-codegen/src/public_api.ts
@@ -2,5 +2,6 @@
  * Public API Surface of angular-codegen
  */
 
+export * from './lib/vue-docgen.service';
 export * from './lib/vue-codegen.service';
 export * from './lib/vue-codegen.module';
