+++ projects/web-component-codegen/src/public_api.ts
@@ -2,5 +2,6 @@
  * Public API Surface of web-component-codegen
  */
 
+export * from './lib/web-component-docgen.service';
 export * from './lib/web-component-codegen.service';
 export * from './lib/web-component-codegen.module';
